fixed overlay close, minimize functionality
